Remove TenancyPrioritiser class
